+++ libs/components/src/lib/button/button.ts
@@ -28,6 +28,7 @@ export type ButtonAppearance = Extract<
 	| Appearance.Outlined
 	| Appearance.Ghost
 	| Appearance.GhostLight
+	| Appearance.OutlinedLight
 >;
 
 /**
@@ -121,6 +122,19 @@ export class Button extends FoundationButton {
 		attribute: 'pending',
 	})
 	pending = false;
+
+	/**
+	 * Display a chevron to indicate that the button opens a dropdown.
+	 *
+	 * @public
+	 * @remarks
+	 * HTML Attribute: dropdown-indicator
+	 */
+	@attr({
+		mode: 'boolean',
+		attribute: 'dropdown-indicator',
+	})
+	dropdownIndicator = false;
 
 	/**
 	 * Displays the button in active state.