+++ libs/components/src/lib/toggletip/ui.test.ts
@@ -10,13 +10,17 @@ const components = ['toggletip'];
 test('should show the component', async ({ page }: { page: Page }) => {
 	const template = `
 	<div style="padding: 20px 100px;">
-		<vwc-button id="button1" icon="help-solid" shape="pill"></vwc-button>
-		<vwc-toggletip open anchor="button1">My anchor is an ID</vwc-toggletip>
+		<vwc-toggletip open anchor="button1">
+		My anchor is an ID
+				<vwc-button slot="anchor" id="button1" icon="help-solid" shape="pill"></vwc-button>
+		</vwc-toggletip>
 	</div>
 
 	<div style="padding: 20px 100px;">
-		<vwc-button id="button3" icon="help-solid" shape="pill"></vwc-button>
-		<vwc-toggletip open anchor="button3" open>I'm open by default</vwc-toggletip>
+		<vwc-toggletip open anchor="button3" open>
+		<vwc-button slot="anchor"  icon="help-solid" shape="pill"></vwc-button>
+		I'm open by default
+		</vwc-toggletip>
 	</div>
 
 	<div style="padding: 20px 100px;">
@@ -40,9 +44,23 @@ test('should show the component', async ({ page }: { page: Page }) => {
 			<vwc-button appearance='filled' label='Action' shape='pill' slot="action-items"></vwc-button>
 		</vwc-toggletip>
 	</div>
+
+	<div style="padding: 50px 100px; display: grid; place-content: center;">
+		<vwc-toggletip open placement="bottom" style="--toggletip-max-inline-size: none">
+			<vwc-button icon="help-solid" slot="anchor" shape="pill"></vwc-button>
+				This is a toggletip with action items and a very long text with no 30ch default max-inline of the toggle tip
+		</vwc-toggletip>
+	</div>
+
+	<div style="padding: 50px 100px; display: grid; place-content: center; block-size: 200px;">
+		<vwc-toggletip open placement="bottom">
+			<vwc-button icon="help-solid" slot="anchor" shape="pill"></vwc-button>
+				This is a toggletip with action items and a very long text with the 30ch default max-inline of the toggle tip
+		</vwc-toggletip>
+	</div>
 	`;
 
-	page.setViewportSize({ width: 500, height: 720 });
+	page.setViewportSize({ width: 720, height: 1300 });
 
 	await loadComponents({
 		page,