+++ libs/wrapper-gen/src/generator/ComponentDef.ts
@@ -8,10 +8,19 @@ export interface ComponentDef {
 	registerFunctionName: string; // Name of the function that registers the component, e.g. registerAccordionItem
 	description?: string;
 	attributes: {
-		name: string; // Attribute on the vivid component in kebab-case. E.g. action-href
+		name: string; // Name of the prop on the Vue component in kebab-case. E.g. action-href
+		forwardTo: // How this prop will be passed to the vivid component
+		| {
+					type: 'attribute';
+					name: string; // Name of the attribute on the vivid component. E.g. action-href
+					boolean?: boolean;
+			  }
+			| {
+					type: 'property';
+					name: string; // Name of the property on the vivid component. E.g. actionHref
+			  };
 		description?: string;
 		type: TypeUnion;
-		forceDomProp?: boolean; // Force the attribute to be set as a DOM property
 	}[];
 	events: {
 		name: string; // Name of the event on the vivid component in kebab-case. E.g. 'change'