+++ libs/components/CHANGELOG.md
@@ -4,6 +4,95 @@ This log was last generated on Thu, 26 Jan 2023 14:40:25 GMT and should not be m
 
 <!-- Start content -->
 
+## [4.3.0](https://github.com/Vonage/vivid-3/compare/vivid-v4.2.0...vivid-v4.3.0) (2024-07-18)
+
+
+### Features
+
+* **button:** add dropdown-indicator option (VIV-1741) ([#1702](https://github.com/Vonage/vivid-3/issues/1702)) ([5caaf84](https://github.com/Vonage/vivid-3/commit/5caaf841352343b147a033f8082d36edfe2a074b))
+* **switch:** add announcement connotation (VIV-1756) ([#1848](https://github.com/Vonage/vivid-3/issues/1848)) ([edf18f8](https://github.com/Vonage/vivid-3/commit/edf18f872b140af806e8da04460fd36879ce4244))
+
+
+### Bug Fixes
+
+* **alert:** adds main slot margin when heading/text is present (VIV-1871) ([#1851](https://github.com/Vonage/vivid-3/issues/1851)) ([1f7cdaa](https://github.com/Vonage/vivid-3/commit/1f7cdaae22acfba2e9a3b554c427aea44dab7637))
+* **fab:** removes elevation (border) when disabled (VIV-1908) ([#1854](https://github.com/Vonage/vivid-3/issues/1854)) ([f5af75d](https://github.com/Vonage/vivid-3/commit/f5af75d22d9f5aef14ca360a52af1d1787d81611))
+* remove elevation from disabled fab ([f5af75d](https://github.com/Vonage/vivid-3/commit/f5af75d22d9f5aef14ca360a52af1d1787d81611))
+
+## [4.2.0](https://github.com/Vonage/vivid-3/compare/vivid-v4.1.0...vivid-v4.2.0) (2024-07-11)
+
+
+### Features
+
+* **menu-item:** emit click event on activation (VIV-1859) ([#1844](https://github.com/Vonage/vivid-3/issues/1844)) ([cc58852](https://github.com/Vonage/vivid-3/commit/cc5885241afda538702b22cef4b5ceffc66973ac))
+
+## [4.1.0](https://github.com/Vonage/vivid-3/compare/vivid-v4.0.0...vivid-v4.1.0) (2024-07-08)
+
+
+### Features
+
+* **alert:** add announcement connotation (VIV-1830) ([#1840](https://github.com/Vonage/vivid-3/issues/1840)) ([b1a980f](https://github.com/Vonage/vivid-3/commit/b1a980fae03f3232b2a7b2d3d15a45c2a024d3a4))
+* **note:** add announcement connotation (VIV-1829) ([#1839](https://github.com/Vonage/vivid-3/issues/1839)) ([9202a3d](https://github.com/Vonage/vivid-3/commit/9202a3d8ae8e69229e95aeccf19874909de93163))
+
+
+### Bug Fixes
+
+* **audio-player:** hide menu when playback rates are not set (VIV-1845) ([#1815](https://github.com/Vonage/vivid-3/issues/1815)) ([f0e32d1](https://github.com/Vonage/vivid-3/commit/f0e32d1584f24f52c3a16ef8a999f52f309543e4))
+* **badge:** centers text inside badge (VIV-1824) ([#1841](https://github.com/Vonage/vivid-3/issues/1841)) ([8be57da](https://github.com/Vonage/vivid-3/commit/8be57da6eef2b433fe001f498d5d76f0b85e5d5c))
+* **number-field:** fix not updating validation after min/max change (VIV-1842) ([5afde52](https://github.com/Vonage/vivid-3/commit/5afde52660594af06e8a18f63facdc312561b4e3))
+* **number-field:** update validation after min/max change (VIV-1842) ([#1833](https://github.com/Vonage/vivid-3/issues/1833)) ([5afde52](https://github.com/Vonage/vivid-3/commit/5afde52660594af06e8a18f63facdc312561b4e3))
+* **select:** fix text color (VIV-1851) ([#1835](https://github.com/Vonage/vivid-3/issues/1835)) ([4e04daa](https://github.com/Vonage/vivid-3/commit/4e04daa157948d13192534b5f0a667dc8b64ff86))
+* **select:** reset form clears selected value if placeholder exists (VIV-1841) ([#1831](https://github.com/Vonage/vivid-3/issues/1831)) ([1faa01b](https://github.com/Vonage/vivid-3/commit/1faa01be71cf220ccb1fc0c6302e86f6e63469fd))
+
+## [4.0.0](https://github.com/Vonage/vivid-3/compare/vivid-v3.60.0...vivid-v4.0.0) (2024-07-01)
+
+
+### ⚠ BREAKING CHANGES
+
+* **tooltip:** sets default max width (VIV-1799) ([#1798](https://github.com/Vonage/vivid-3/issues/1798))
+* **data-grid-cell:** change default value of block-size and white-space (VIV-1797) ([#1793](https://github.com/Vonage/vivid-3/issues/1793))
+* **audio-player:** remove css-variable for min-inline-size and max-inline-size on host (VIV-1800) ([#1803](https://github.com/Vonage/vivid-3/issues/1803))
+* **vue-wrappers:** forward form values by DOM prop name (VIV-1798) ([#1789](https://github.com/Vonage/vivid-3/issues/1789))
+* **vue-wrappers:** forward form values by DOM prop name (VIV-1798)
+* **tabs:** change gutters default to small (VIV-1796) ([#1804](https://github.com/Vonage/vivid-3/issues/1804))
+* **docs:** refine docs for v-4 (VIV-000) ([#1813](https://github.com/Vonage/vivid-3/issues/1813))
+* **toggletip:** add max-inline-size of 30ch & css-variable for custom size (VIV-1846) ([#1814](https://github.com/Vonage/vivid-3/issues/1814))
+
+### Features
+
+* **audio-player:** remove css-variable for min-inline-size and max-inline-size on host (VIV-1800) ([#1803](https://github.com/Vonage/vivid-3/issues/1803)) ([06ee68a](https://github.com/Vonage/vivid-3/commit/06ee68a6918612569f61a866716c1b1b96f1d1d6))
+* **data-grid-cell:** change default value of block-size and white-space (VIV-1797) ([#1793](https://github.com/Vonage/vivid-3/issues/1793)) ([06ee68a](https://github.com/Vonage/vivid-3/commit/06ee68a6918612569f61a866716c1b1b96f1d1d6))
+* **toggletip:** add max-inline-size of 30ch & css-variable for custom size (VIV-1846) ([#1814](https://github.com/Vonage/vivid-3/issues/1814)) ([06ee68a](https://github.com/Vonage/vivid-3/commit/06ee68a6918612569f61a866716c1b1b96f1d1d6))
+* **tooltip:** sets default max width (VIV-1799) ([#1798](https://github.com/Vonage/vivid-3/issues/1798)) ([06ee68a](https://github.com/Vonage/vivid-3/commit/06ee68a6918612569f61a866716c1b1b96f1d1d6))
+* **vue-wrappers:** forward form values by DOM prop name (VIV-1798) ([06ee68a](https://github.com/Vonage/vivid-3/commit/06ee68a6918612569f61a866716c1b1b96f1d1d6))
+* **vue-wrappers:** forward form values by DOM prop name (VIV-1798) ([#1789](https://github.com/Vonage/vivid-3/issues/1789)) ([06ee68a](https://github.com/Vonage/vivid-3/commit/06ee68a6918612569f61a866716c1b1b96f1d1d6))
+
+
+### Bug Fixes
+
+* **tabs:** change gutters default to small (VIV-1796) ([#1804](https://github.com/Vonage/vivid-3/issues/1804)) ([06ee68a](https://github.com/Vonage/vivid-3/commit/06ee68a6918612569f61a866716c1b1b96f1d1d6))
+
+
+### Miscellaneous Chores
+
+* **docs:** refine docs for v-4 (VIV-000) ([#1813](https://github.com/Vonage/vivid-3/issues/1813)) ([06ee68a](https://github.com/Vonage/vivid-3/commit/06ee68a6918612569f61a866716c1b1b96f1d1d6))
+
+## [3.60.0](https://github.com/Vonage/vivid-3/compare/vivid-v3.59.0...vivid-v3.60.0) (2024-07-01)
+
+
+### Features
+
+* **buttons:** add outline-light appearance (VIV-1775) ([#1791](https://github.com/Vonage/vivid-3/issues/1791)) ([6eebefe](https://github.com/Vonage/vivid-3/commit/6eebefe99566c5df57546757d822453d6ca3264c))
+
+
+### Bug Fixes
+
+* **audio-player:** keep paused state while dragging (VIV-1787) ([#1792](https://github.com/Vonage/vivid-3/issues/1792)) ([90d7a58](https://github.com/Vonage/vivid-3/commit/90d7a58667bc68802a152ce53459513d0297783a))
+* disabled buttons in number field ([219bba7](https://github.com/Vonage/vivid-3/commit/219bba7a6797c13854d10e486ae2e57c8574db89))
+* **number-field:** disabled buttons in number field (VIV-1817) ([#1779](https://github.com/Vonage/vivid-3/issues/1779)) ([219bba7](https://github.com/Vonage/vivid-3/commit/219bba7a6797c13854d10e486ae2e57c8574db89))
+* **select:** fix initial select validity (VIV-1822) ([#1794](https://github.com/Vonage/vivid-3/issues/1794)) ([e8ea809](https://github.com/Vonage/vivid-3/commit/e8ea809fc047b43e5beccfd615ae51d9ee368d59))
+* **text-area:** row="1" min-height (VIV-1647) ([#1783](https://github.com/Vonage/vivid-3/issues/1783)) ([db7277a](https://github.com/Vonage/vivid-3/commit/db7277afc00bf38bb433464ef0785b543a7dc28d))
+
 ## [3.59.0](https://github.com/Vonage/vivid-3/compare/vivid-v3.58.0...vivid-v3.59.0) (2024-06-18)