+++ libs/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [1.2.1](https://github.com/Vonage/vivid-3/compare/eslint-plugin-v1.2.0...eslint-plugin-v1.2.1) (2024-07-08)
+
+
+### Bug Fixes
+
+* **eslint-plugin:** add missing export of accessible-names (VIV-000) ([#1838](https://github.com/Vonage/vivid-3/issues/1838)) ([bd53694](https://github.com/Vonage/vivid-3/commit/bd5369414f5ae657b5e87118bb10d0622a6a8450))
+
+## [1.2.0](https://github.com/Vonage/vivid-3/compare/eslint-plugin-v1.1.0...eslint-plugin-v1.2.0) (2024-07-01)
+
+
+### Features
+
+* **eslint-plugin:** add accessible names rule (VIV-1698) ([#1806](https://github.com/Vonage/vivid-3/issues/1806)) ([e06e911](https://github.com/Vonage/vivid-3/commit/e06e911bbad87ef9e6d69b7b03f6859086566e16))
+* **eslint-plugin:** add no value attribute rules (VIV-1827) ([#1785](https://github.com/Vonage/vivid-3/issues/1785)) ([6a3e495](https://github.com/Vonage/vivid-3/commit/6a3e495896c91a079c1f524bea258a7d6fc12ec1))
+* **eslint-plugin:** add no-anchor-attribute rule (VIV-1701) ([#1752](https://github.com/Vonage/vivid-3/issues/1752)) ([94dd40c](https://github.com/Vonage/vivid-3/commit/94dd40c741b919eae34d04c6dd06a40c7b6b49fc))
+
 ## [1.1.0](https://github.com/Vonage/vivid-3/compare/eslint-plugin-v1.0.0...eslint-plugin-v1.1.0) (2024-06-18)